+++ util/fileutil/fileutil_test.go
@@ -1,0 +1,13 @@
+package fileutil
+
+import "testing"
+
+func TestRead(t *testing.T) {
+
+	Read(".")
+
+	got := 2
+	if got != 1 {
+		t.Errorf("Abs(-1) = %d; want 1", got)
+	}
+}
